+++ src/PJM/AppBundle/Controller/API/PushSubscriptionController.php
@@ -1,70 +1,32 @@
 <?php
 
-namespace PJM\AppBundle\Controller;
+namespace PJM\AppBundle\Controller\API;
 
-use Symfony\Bundle\FrameworkBundle\Controller\Controller;
-use Symfony\Component\HttpFoundation\Request;
-use Symfony\Component\HttpFoundation\JsonResponse;
-use Symfony\Component\HttpFoundation\Response;
 use PJM\AppBundle\Entity\PushSubscription;
+use Sensio\Bundle\FrameworkExtraBundle\Configuration\Method;
+use Sensio\Bundle\FrameworkExtraBundle\Configuration\Route;
+use Sensio\Bundle\FrameworkExtraBundle\Configuration\Security;
+use Symfony\Bundle\FrameworkBundle\Controller\Controller;
+use Symfony\Component\HttpFoundation\JsonResponse;
+use Symfony\Component\HttpFoundation\Request;
+use Symfony\Component\HttpFoundation\Response;
 
-class PushController extends Controller
+/**
+ * @Route("/pushSubscriptions")
+ * @Security("has_role('ROLE_USER')")
+ */
+class PushSubscriptionController extends Controller
 {
     /**
-     * Action ajax de rendu de la liste des pushSubscriptions.
+     * @param Request $request
+     * @param bool|false $action
+     * @return JsonResponse|\Symfony\Component\HttpFoundation\RedirectResponse
+     *
+     * @Route("/manage/{action}", options={"expose"=true})
+     * @Method("POST")
      */
-    public function subscriptionResultsAction()
+    public function manageAction(Request $request, $action)
     {
-        $datatable = $this->get('pjm.datatable.pushsubscription');
-        $datatable->buildDatatable();
-
-        $query = $this->get('sg_datatables.query')->getQueryFrom($datatable);
-
-        $em = $this->getDoctrine()->getManager();
-        $repository = $em->getRepository('PJMAppBundle:PushSubscription');
-        $query->addWhereAll($repository->callbackFindByUser($this->getUser()));
-
-        return $query->getResponse();
-    }
-
-    /**
-     * Action ajax de suppression de pushSubscription.
-     */
-    public function deleteSubscriptionAction(Request $request)
-    {
-        if ($request->isXmlHttpRequest()) {
-            $liste = $request->request->get('data');
-
-            $em = $this->getDoctrine()->getManager();
-            $repository = $em->getRepository('PJMAppBundle:PushSubscription');
-
-            foreach ($liste as $choice) {
-                $pushSubscription = $repository->find($choice['value']);
-
-                if ($pushSubscription->getUser() != $this->getUser()) {
-                    return new Response("La PushSubscription ne correspond pas à l'utilisateur.", 403);
-                }
-
-                $em->remove($pushSubscription);
-            }
-
-            $em->flush();
-
-            return new Response('PushSubscription(s) removed.');
-        }
-
-        return $this->redirect($this->generateUrl('pjm_app_homepage'));
-    }
-
-    /**
-     * Action ajax de gestion d'une PushSubscription.
-     */
-    public function manageSubscriptionAction(Request $request, $action = false)
-    {
-        if (!$request->isXmlHttpRequest()) {
-            return $this->redirect($this->generateUrl('pjm_app_homepage'));
-        }
-
         $annuler = ($action == 'annuler') ? true : false;
         $subscription = array(
             'id' => $request->request->get('id'),
@@ -125,5 +87,55 @@ class PushController extends Controller
             'done' => $action,
             'subscription' => $subscription,
         ));
+    }
+
+    /**
+     * (DataTable) Results.
+     *
+     * @return Response
+     *
+     * @Route("/results")
+     * @Method("GET")
+     */
+    public function resultsAction()
+    {
+        $datatable = $this->get('pjm.datatable.pushsubscription');
+        $datatable->buildDatatable();
+        $query = $this->get('sg_datatables.query')->getQueryFrom($datatable);
+        $repository = $this->getDoctrine()->getManager()->getRepository('PJMAppBundle:PushSubscription');
+        $query->addWhereAll($repository->callbackFindByUser($this->getUser()));
+        return $query->getResponse();
+    }
+
+
+    /**
+     * (DataTable) Delete choices.
+     *
+     * @param Request $request
+     * @return Response
+     *
+     * @Route("/delete", options={"expose"=true})
+     * @Method("POST")
+     */
+    public function deleteAction(Request $request)
+    {
+        $list = $request->request->get('data');
+
+        $em = $this->getDoctrine()->getManager();
+        $repository = $em->getRepository('PJMAppBundle:PushSubscription');
+
+        foreach ($list as $choice) {
+            $pushSubscription = $repository->find($choice['value']);
+
+            if ($pushSubscription->getUser() != $this->getUser()) {
+                throw $this->createAccessDeniedException();
+            }
+
+            $em->remove($pushSubscription);
+        }
+
+        $em->flush();
+
+        return new Response('PushSubscription(s) removed.');
     }
 }
