--- src/PJM/AppBundle/Resources/config/Push/routing.yml
@@ -1,19 +1,0 @@
-pjm_app_push_subscriptionResults:
-    path:     /subscription/results
-    methods: [ get, post ]
-    defaults: { _controller: PJMAppBundle:Push:subscriptionResults }
-
-pjm_app_push_manageSubscription:
-    path:     /subscription/manage/{action}
-    methods: [ get, post ]
-    defaults: { _controller: PJMAppBundle:Push:manageSubscription, action: false }
-    options:
-        expose: true
-
-pjm_app_push_deleteSubscription:
-    path:     /subscription/delete
-    methods: [ get, post ]
-    defaults: { _controller: PJMAppBundle:Push:deleteSubscription }
-    options:
-        expose: true
-
